[helm-toolkit] Update toolkit to support fqdn alias
This change add the ability to add fqdn alias to namespace and cluster ingress resources. This change is specifically required for keystone so HA of backup solution can be implemented.This change allows user to specify alias_fqdn in the endpoints section, and user can have alias configued. This change is backward compatible, so without specifying this option in charts gives one fqdn ingress rule without cname alias as default behaviour.

Change-Id: Ib1c60524e2f247bb057318b1143bfbc3bde5b73a