[glance] Add override for downloading Ubuntu image
We used to add the Ubuntu image to the list of
images to download during the glance bootstrap
in one of our test script. This image is mostly
used for debugging and when cirros is incompatible
with infra VMs. So let's add glance values override
file specifically for this and enable it only when
necessary.

Change-Id: I5c10dac4826c3ea10327e4934bbeab27b54877e2