Run ovn controller with non root openvswitch user
We recently updated the openvswitch chart to run
ovs db server as non root.

See: https://review.opendev.org/c/openstack/openstack-helm-infra/+/939580

Also ovn-kubernetes script ovnkube.sh that we are using for
lifecycle management of OVN components tries to update the
ownership of OVS run and config directories before start.

So we have to pass the correct username to the script
so it does not break the OVS files permissions.

Change-Id: Ie00dd2657c616645ec237c0880bbc552b3805236