[k8s,ceph,docker] Apt repository filename cleanup
The prerequisites and containerd tasks that add the Kubernetes,
Ceph, and Docker repos to apt list the filenames as
kubernetes.list, ceph.list, and docker.list, which results in the
files created under /etc/apt/sources.list.d being named with
'.list.list' extensions. This change simply removes the redundant
'.list' from the filenames to clean that up.

Change-Id: I3672873149d137ad89c176cabad4c64dcff2bfee