helm-toolkit: always add pod mounts for db-sync job
Always include mounts defined for the db-sync job under the pods section
rather than requiring every chart to pass it in explicitly. Now the
passed in value can be just for overrides. Since charts today already
pass this in, we need to de-duplicate it to ensure we don't create this
multiple times.

Change-Id: I042e79cee7859ebdc001a056edc75eb89dd3e5b3