Cleanup metallb feature gate
metallb isn't really a feature gate because
it is not related to any override. Instead we
used it only during ingress deployment to decide
whether we need to deploy cluster ingress instance.
Previously we moved ingress deployment to the deploy-env
role and cluster ingress instance is deployed
when metallb_setup job variable is set.

Change-Id: I22f6864e40f86527826a40379e516228f43896da