Update ceph-rook.sh script
While we are waiting for Ceph cluster to be ready
we check Ceph status in a loop using tools pod provided by Rook.
We have to get this tools pod name every iteration
within the loop because K8s can terminate the pod for
some reason and this is expected behavior.

Change-Id: Iabb98e94d7470fe996091bf77787637f3e8f4798