[ceph] Upgrade Ceph and Rook
Ceph 19.2.1 contains a critical RGW bug that can lose data. 19.2.2
is a hotfix release to address that issue.

This change updates the Ceph charts to new Ceph 19.2.2 images that
also contain a patch for another RGW bug that was previously
patched onto 19.2.1.

The Rook deployment is also updated to Rook 1.16.6 and is
configured to deploy Ceph 19.2.2.

Change-Id: I05ec9ed447e43a5b36f3004a84eb53cc846bc537