Update versions of all charts to 2025.1.0
Also update chart_version.sh script which
counts the number of commits since the BASE_VERSION
if BASE_VERSION tag is found. If the tag
doesn't exist it counts the number of commits since the
beginning. So when we use for example 2025.1.0 as
the BASE_VERSION but the tag 2025.1.0 is not yet created
the version of the e.g. nova chart will be calculated as
like 2025.1.563+<sha> and then when we create tag the nova version
will be 2025.1.0+<sha> which is undesired.

Lets use BASE_VERSION-<sha> if the tag is not found.

Change-Id: I032e8269ab17b490898d190adaec5c282e96fa88