Add interface name parameter for DPDK configs
Implemented parameter 'iface' in the nics dictionaries which allows
users to reference the interfaces by name instead of only 'pci_id' in
the values.yaml.  This can ease the configuration of interfaces that may
have differing PCI addresses among the servers. 'pci_id' will take
precedence if both are configured.

Change-Id: I93248bb580460c48fff01130d80e2bfe31dadf72