Use Metallb for all jobs by default
The current way how we utilize K8s Ingress API
is complicated and we agreed to re-design it.

It assumes we have two ingress controller instances:
- cluster wide instance
- namespace instance

Cluster wide instance is deployed in the host network namespace
and listens on node IPs. We can use it to get access to services
running inside the k8s cluster if we resolve names into cluster IPs.

Namespace instance listens on cluster IPs and is only
available inside the cluster or via a load balancer like Metallb.
For tests we utilize this instance together with the cluster.local
domain suffix to get access to services running on top of K8s.
However cluster.local is not intended to be used outside the cluster.
The *.cluster.local names are not supposed to be visible outside
and should be used only for the communication inside the cluster.

Load balancers or other more generally gateway instances (see Gateway API)
should be used to get access to services running in K8s clusters.

This PS is a step towards clearer ingress implementation.

Change-Id: I57bee6e0f82c9deb2745e8e0d18c420b74837421