[nova]: split service accounts from the main nova.conf
Split out the service account generation from the main nova.conf and
into their own snippets which get mounted into the pods. OSLO reads
these snippets and appends them to loaded configuration. This is a
follow on to the spec [1] which has separate service accounts for
accessing different services. The goal of separating this out is to
allow users to provide these secrets via other means like Vault or
External Secrets.

[1] I12eb9341d5ff633ad4435f4938bf8c946ea388ee

Change-Id: Ic05ed2886ce446803975124b755cbabe010680b6
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>