Adds support to UUID deployment-provisioning configuration
This change adds new fields in Nova's values.yaml file to allow the
choice of deployment-provisioning the compute node's UUID.
This is done by including two options into the file. The first one,
'hosts_uuids', is a list containing the hostnames and uuids of the
compute nodes where nova-compute pods are being deployed. Each node
is identified by two key-value pairs (name: <hostname> and uuid:
<uuid>). It defaults to an empty list and can be populated via
overrides.
The next option is a boolean value indicating if self provisioning
should be enabled/disabled, 'manifests.compute_uuid_self_provisioning'.
It defaults to true, meaning that nova-compute will provision the
compute node's UUID. When set to false, the deployment-provisioning
is enabled.
This change also adds a code in nova-compute-init.sh that searches for
the current node's UUID and copies it to the directory indicated by
the 'state_path' field in nova.conf file, if self-provision is disabled.
By doing this, we are provisioning the UUID, accordingly to [1].

[1] https://docs.openstack.org/nova/latest/admin/compute-node-identification.html

Change-Id: I3101e5bcdd6acf2b950e196e9b54d5f900501e68
Signed-off-by: Joao Fracarolli <joao.vicentinifracarolli@windriver.com>