fix(cinder): provide a pod writeable space for state_path
Some cinder volume backends will need to write data or mount and they
will always do so relative to the state_path. This defaults the
state_path to the same value in the cinder.conf and provides an emptyDir
mount so that the path is writeable inside the pod. This is the same
behavior done in the neutron chart.

Change-Id: Iaa319e4b5f3aa1a18970f2f699804606301b0c92
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>