chore: add pre-commit and hook it into Zuul
A number of OpenStack projects have adopted pre-commit and run it as
their lint step so that developers have an easy way to run the same
checks locally that will run in the CI loop. So this adds an initial
pre-commit config to detect trailing whitespace and replaces that
existing check to be called via pre-commit. Added a tox environment for
running pre-commit as well.

Change-Id: I9cf95f0c6db166ee1e59f8aedf376c511ac23010
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>