[ceph-mgr] Remove grep from the ceph-mgr health check function
Ceph 19.2.3 replaces the handler for the ceph-mgr asok status query
with a new handler that simply returns an empty dictionary and a
success code. The ceph-mgr readiness/liveness probe health check
function currently relies on content from the returned dictionary
to succeed. This change simply removes that content check and
relies solely on the return code of the status query instead.

This change to the ceph-mgr probes retains backward compatiblity
with previous Ceph versions as well.

Change-Id: Iffa1764735733a2b2ef16c5cba6c9880d03b357b
Signed-off-by: Stephen Taylor <stephen.taylor.1@att.com>