fix(cinder): provide writable state_path for cinder-scheduler
Some cinder services (including the scheduler) require a writable
state_path (typically /var/lib/cinder) for coordination and runtime
files. Previously, the cinder-scheduler pod was missing a writable
mount for this directory, causing the service to terminate after
initialization.

This patch aligns the scheduler deployment with the cinder-volume
pattern by defining an emptyDir volume for the state_path and mounting
it into the container. The cinder.conf `state_path` is also set to the
same value for consistency.

Change-Id: I637f344b7ce361213d7bcda4f1691e35797c811e
Signed-off-by: Saeed Padari <sam137115@gmail.com>