ironic: drop the retrieve-cleaning-network init container
It was only setting a deprecated option and the script it ran gave less
feedback to the user than the built in check inside of Ironic. With the
future of Ironic standalone networking this option will make even less
sense so allow the default behavior of looking up the name in Ironic.

Change-Id: Iee328fb7f7ed62beacb7471c23f899849b44e300
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>