feat: build charts into sub dirs
As outlined in
https://lists.opendev.org/archives/list/service-discuss@lists.opendev.org/thread/VTMDDVSPM5HRUYWAATNMZOILT5OE57VR/
the current structure of building all the charts into one directory is
causing issues on the opendev infra due too many entries in one
directory. Switch away from using a Makefile to using an Ansible role to
build each chart and then use chart-testing to identify the charts that
need to be rebuilt and lastly build them build put the output into a
subdir matching the chart name.

Change-Id: I61f11950ba381c7897eb6bfff05a508ca4db9f06
Signed-off-by: Doug Goldstein <cardoe@cardoe.com>
Signed-off-by: Vladimir Kozhukalov <kozhukalov@gmail.com>